requirements.txt
# Use PyTorch with CUDA 12.1 support
--index-url https://download.pytorch.org/whl/cu128
--extra-index-url https://pypi.org/simple

# Core PyTorch stack
torch>=2.0.0
torchvision
torchaudio

# Hugging Face ecosystem
transformers>=4.30.0
peft>=0.4.0
datasets>=2.12.0
accelerate>=0.20.0
sentencepiece>=0.1.99

# Optimization and quantization
bitsandbytes>=0.39.0

# Experiment tracking
wandb>=0.15.0

# Data and utilities
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
tqdm>=4.65.0
pyyaml>=6.0
protobuf>=3.20.0
requests>=2.31.0

# Web scraping and parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0
